Handle incorrect coordinates give by piler
The coordinates that PILER-CR reports are incorrect if the repeat sequence contains gaps. Each start coordinate is off by the number of gaps in all preceding repeat-spacers. The parser now adjusts all coordinates to what they should have been.

The author of PILER-CR says that the software is no longer being maintained and that this bug won't be fixed, so this parser should always be compliant with the output.